feat: 피드 숨기기 API 분리
