+++ src/app/mocks/handler/feed.ts
@@ -189,22 +189,4 @@ export const feedHandlers = [
 
     return createHttpSuccessResponse({});
   }),
-
-  // 7️⃣ 피드 숨기기
-  http.put('/feeds/:feed_id/hides', ({ params }) => {
-    const { feed_id } = params;
-
-    if (isNaN(Number(feed_id))) {
-      return createHttpErrorResponse('4220');
-    }
-
-    const formattedFeedId = Number(feed_id);
-
-    if (!feeds[formattedFeedId]) {
-      return createHttpErrorResponse('4040');
-    }
-
-    hiddens[formattedFeedId] = true;
-    return createHttpSuccessResponse({ isHidden: true });
-  }),
 ];
